WIP migration off ReactMapGL
