+++ src/TopoMap.tsx
@@ -18,10 +18,43 @@ import Terrain from './layers/terrain';
 import SettingsSection from './sections/SettingsSection';
 import MenuSection from './sections/MenuSection';
 import MenuControl from './controls/MenuControl';
-import JMap from './map/Map'
+import JMap, { useMap } from './map/Map'
+import Layer from './map/Layer';
+import Source from './map/Source';
+
+const sources = baseLayers.flatMap(b => Object.entries(b.sources).map(([key, spec]) => <Source key={key} id={key} spec={spec as any} />))
+const terrain = {
+    source: 'dem',
+    exaggeration: 1
+}
+function Layers() {
+
+    const routeParams = useParams();
+    const { map } = useMap()
+    const basemap = baseLayers.find(r => r.id === routeParams.basemap) ?? linzVector
+
+    // Swap to/from 3d mode when the view changes
+    React.useEffect(() => {
+        const t = routeParams.pitch === 0 ? null : terrain;
+        if (map?.getTerrain() !== t) {
+            map.on('idle', () => {
+                map?.setTerrain(t)
+            })
+        }
+    }, [routeParams.pitch])
+
+    return <>
+        <Terrain />
+        {basemap.layers.map(l => <Layer key={l.id} layer={l as any} />)}
+        {overlays.filter(e => routeParams.overlays.includes(e.id)).map(o => typeof o.source === 'function' ? <o.source key={o.id} /> : o.source)}
+    </>
+}
 
 export default function TopoMap() {
+
     return <JMap>
+        {sources}
+
         <SearchSection />
         <MountainsSection />
         <MountainSection />
@@ -32,6 +65,8 @@ export default function TopoMap() {
         <LayersControl />
         <MenuControl />
         <SearchControl />
+
+        <Layers />
     </JMap>
     // return <Map
     //     ref={mapRef as any}
